rename forecast-enrollment to newsletter
